pyproject.toml
[tool.poetry]
name = "shuncommands"
version = "1.0.3"
description = "mypaceshun usefull commands"
authors = ["KAWAI Shun <mypaceshun@gmail.com>"]
maintainers = ["KAWAI Shun <mypaceshun@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mypaceshun/shuncommands"
repository = "https://github.com/mypaceshun/shuncommands"
keywords = ["python", "click"]
packages = [
  {include = 'shuncommands/'}
]
include = ["CHANGELOG.md"]

[tool.poetry-dynamic-versioning]
enable = true
vcs = 'git'
style = 'pep440'

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0"
gitignore-parser = "^0.0.8"
pikepdf = "^4.1.0"
"pdfminer.six" = "^20211012"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
poethepoet = "^0.11.0"
autoflake = "^1.4"
isort = "^5.10.1"
git-changelog = "^0.5.0"
pytest = "^6.2.5"
black = "^21.12b0"
pyproject-flake8 = "^0.0.1-alpha.2"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"

[tool.poetry.scripts]
rmtmp = "shuncommands.rmtmp:ctx"
pdftool = "shuncommands.pdftool:ctx"

[tool.poe.tasks.changelog]
cmd = "git-changelog -o CHANGELOG.md .git"

[tool.poe.tasks.lint]
cmd = "pflake8 shuncommands/ tests/"

[tool.poe.tasks.format]
sequence = [
  { cmd = "autoflake -ir --remove-all-unused-imports --remove-unused-variables --ignore-init-module-imports shuncommands/ tests/" },
  { cmd = "isort --profile black shuncommands/ tests/" },
  { cmd = "black shuncommands/ tests/" }
]
[tool.poe.tasks.test]
cmd = "pytest --cov tests/"

[tool.flake8]
max-line-length = 88
ignore = "E203,W503,W504"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
